backend/requirements.txt
asgiref==3.4.1
astroid==2.8.3
autopep8==1.5.7
beautifulsoup4==4.10.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
chrome-gnome-shell==0.0.0
colorama==0.4.4
cryptography==35.0.0
cssselect2==0.5.0
cupshelpers==1.0
dbus-python==1.2.18
defusedxml==0.7.1
distlib==0.3.4
distro==1.7.0
distro-info==1.1
Django==3.2.8
django-allauth==0.45.0
django-autotranslate==1.2.0
django-cors-headers==3.10.0
django-extensions==3.1.5
django-filter==21.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
dnspython==2.2.0
dtrx==8.2.2
ez-setup==0.9
filelock==3.4.2
fpdf2==2.4.6
futures==3.0.5
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
goslate==1.5.2
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
isort==5.9.3
lazy-object-proxy==1.7.1
lxml==4.7.1
Markdown==3.3.4
mccabe==0.6.1
mysqlclient==2.1.0
numpy==1.21.5
oauthlib==3.1.1
pep8==1.7.1
Pillow==8.4.0
platformdirs==2.4.0
polib==1.1.1
protobuf==3.12.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycodestyle==2.8.0
pycparser==2.20
pycups==2.0.1
pycurl==7.44.1
pydotplus==2.0.2
PyGObject==3.42.0
PyJWT==2.3.0
pylint==2.11.1
PyMySQL==1.0.2
pyparsing==2.4.7
PyPDF2==1.26.0
PySimpleSOAP==1.16.2
pysmbc==1.0.23
python-apt==2.3.0+b1
python-debian==0.1.43
python-debianbts==3.2.0
python-xlib==0.29
python3-openid==3.2.0
pytz==2021.3
pyudev==0.22.0
pyvips==2.1.16
PyYAML==5.4.1
reportbug==11.4.1
reportlab==3.6.6
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.8
scour==0.38.2
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
toml==0.10.2
typing-extensions==3.10.0.2
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.13.1
webencodings==0.5.1
wrapt==1.13.2
